Delete vcard - DONE

+++ laravelAPI/app/Http/Controllers/VcardController.php
@@ -10,6 +10,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests\VcardPut;
 use App\Models\DefaultCategory;
 use App\Http\Requests\VcardPost;
+use App\Http\Requests\VcardDelete;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Resources\VcardResource;
@@ -73,8 +74,39 @@ class VcardController extends Controller
         return new VcardResource($vcard);
     }
 
-    public function destroyVcard(Vcard $vcard){
-        $vcard->delete();
+    public function destroyVcard(VcardDelete $request, Vcard $vcard){
+
+        if(Hash::check($vcard->password, $request->password)){
+            throw ValidationException::withMessages(['password' => "Password is not correct"]);
+        }
+        if(Hash::check($vcard->confirmation_code, $request->confirmation_code)){
+            throw ValidationException::withMessages(['confirmation_code' => "PIN is not correct"]);
+        }
+        if($vcard->balance > 0){
+            throw ValidationException::withMessages(['balance' => "Vcard cannot be deleted - Balance is bigger than 0.00"]);
+        }
+        $transactions = $vcard->transactions;
+        $numberTransactions = count($transactions);
+        try{
+            if($numberTransactions > 0){
+                foreach($transactions as $transaction){
+                    $transaction->delete();
+                }
+                $vcard->delete();
+            }
+            if($numberTransactions == 0){
+                $categories = $vcard->categories;
+                foreach($categories as $category){
+                    $category->forceDelete();
+                }
+                $vcard->forceDelete();
+            }
+
+        }catch(Exception $e){
+            DB::rollback();
+            throw new Exception("Error deleting the vcard");
+        }
+
         return new VcardResource($vcard);
     }
